requirements.txt
nltk
colorama
pyfiglet
pyenchant